Fix index-out-of-range bug
Under certain circumstances, orderSearchResults grouping the SQL result
set failed due to a wrong index that was used to store the tracks
relevance orders determined by the database. More specifically, the
original index position of the result set array was used and therefore
sometimes exceeded the length of the results array.